backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.27.1
websockets==12.0

# LangGraph & AI
langgraph==0.3.27
langchain-core>=0.3.23,<0.4.0
langchain-google-genai==2.0.4
langchain-community>=0.0.40

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
chromadb==0.4.15
redis==5.0.1

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.9.0

# ML & NLP
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Web Scraping
selenium==4.15.0
beautifulsoup4==4.12.2
requests==2.31.0

# Email & Calendar
# smtplib
# google-api-python-client==2.100.0

# Utilities
python-dotenv==1.0.1
pydantic
python-multipart==0.0.6


google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0

urllib3==2.0.7
jinja2==3.1.2